Allow to define mime-type when uploading a new file (#58)
* Skip PUT Upload test

This skips a failing test that depends on an httpbin feature that
is currently not working.
Ref: https://github.com/kennethreitz/httpbin/issues/340#issuecomment-330176449

* Add support to custom mimetypes in file uploads

This adds a new FileMime field to the FileUpload struct. When
present, this field will be used to provide the Content-Type of
a field being uploaded. Otherwise, the http package will be
responsible for providing the value. Currently, the mentioned
package will default to application/octet-stream